requirements.txt
torch>=2.0.0
transformers>=4.30.0
gradio>=6.0.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0